Gym version fix nightly

+++ tests/testing_utils.py
@@ -150,9 +150,9 @@ def mock_instantiate(superclass, key, **kwargs):
 class DummyReferenceGenerator(ReferenceGenerator):
     _reset_counter = 0
 
-    def __init__(self, reference_observation=np.array([1]), reference_state='dummy_state_0', **kwargs):
+    def __init__(self, reference_observation=np.array([1.], dtype=np.float32), reference_state='dummy_state_0', **kwargs):
         super().__init__()
-        self.reference_space = Box(0, 1, shape=(1,))
+        self.reference_space = Box(0, 1, shape=(1,), dtype=np.float64)
         self.kwargs = kwargs
         self._reference_names = [reference_state]
         self.closed = False
@@ -257,7 +257,7 @@ class DummyPhysicalSystem(PhysicalSystem):
         self.kwargs = kwargs
 
     def reset(self, initial_state=None):
-        self.state = np.array([0] * len(self._state_names))
+        self.state = np.array([0.] * len(self._state_names), dtype=np.float32)
         return self.state
 
     def simulate(self, action):
